Clean up duplicate webapp splits and consolidate to single entry point
- Remove app_core.py (duplicate initialization)
- Remove split_web_app_complete.py (old approach)
- Remove refactor_web_app.py (old approach)
- Remove unused routes/ directory
- Fix routes_main.py to define admin_required locally instead of importing from app_core
- Consolidate all Flask app logic into web_app.py as the single entry point
- Keep clean separation: web_app.py (initialization & blueprints) + routes_*.py (feature routes)
