--- refactor_web_app.py
@@ -1,88 +1,0 @@
-#!/usr/bin/env python3
-"""
-Script to refactor web_app.py into 3 files with PostgreSQL-only code
-"""
-
-# Read the original file
-with open('web_app.py', 'r') as f:
-    lines = f.readlines()
-
-# Define line ranges for each section (approximate - will refine)
-# Based on grep output:
-# Auth routes: 274-475
-# Main routes: 481-496
-# Drafts/Listings: 498-545
-# Storage routes: 550-724
-# Settings: 726-768
-# Admin: 771-910
-# API endpoints: 915-1920
-# Settings API: 1795-1920
-# Baby bird: 1924-1963
-# Cards: 1965-2494
-
-# Let's first identify all the imports and helper functions needed
-
-header = """#!/usr/bin/env python3
-\"""
-AI Cross-Poster Web App - Main Entry Point
-============================================
-Mobile-friendly web interface for inventory management and cross-platform listing.
-
-PostgreSQL-compatible version - split into modular routes.
-\"""
-
-import os
-from pathlib import Path
-from flask import Flask
-from flask_login import LoginManager
-from werkzeug.security import generate_password_hash
-from dotenv import load_dotenv
-
-from src.database import get_db
-
-# Load environment
-load_dotenv()
-
-# Initialize Flask app
-app = Flask(__name__)
-app.secret_key = os.getenv('FLASK_SECRET_KEY', 'dev-secret-key-change-in-production')
-app.config['MAX_CONTENT_LENGTH'] = 16 * 1024 * 1024  # 16MB max upload
-app.config['UPLOAD_FOLDER'] = './data/uploads'
-
-# Ensure upload folder exists
-Path(app.config['UPLOAD_FOLDER']).mkdir(parents=True, exist_ok=True)
-
-# Initialize database
-db = get_db()
-
-# Initialize Flask-Login
-login_manager = LoginManager()
-login_manager.init_app(app)
-login_manager.login_view = 'auth.login'
-login_manager.login_message = 'Please log in to access this page.'
-"""
-
-print("Refactoring web_app.py...")
-print("This script will create 3 new files and backup the original.")
-print("\nCreating backup...")
-
-# Create backup
-import shutil
-shutil.copy('web_app.py', 'web_app.py.backup')
-print("✓ Backup created: web_app.py.backup")
-print("\nAnalyzing file structure...")
-print(f"Total lines: {len(lines)}")
-
-# Count routes
-import re
-routes = [i for i, line in enumerate(lines) if re.match(r'^@app\.route', line)]
-print(f"Total routes found: {len(routes)}")
-
-# Print route list with line numbers
-print("\nRoute breakdown:")
-for i in routes[:10]:
-    print(f"  Line {i+1}: {lines[i].strip()}")
-print(f"  ... and {len(routes)-10} more routes")
-
-print("\nDue to file complexity, creating simplified structure...")
-print("You'll need to manually verify and adjust the splits.")
