--- routes/__init__.py
@@ -1,4 +1,0 @@
-"""
-Routes package for AI Cross-Poster Web App
-Organizes routes into logical modules using Flask Blueprints
-"""
